Renamed `api/predictions` to `api/predicationApi`
